tests/wpt/web-platform-tests/fetch/metadata/sec-fetch-dest/img.tentative.https.sub.html

<!DOCTYPE html>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=/common/security-features/resources/common.sub.js></script>
<script src=/fetch/metadata/resources/helper.js></script>
<body>
<script>
  // These tests reuse the `referrer-policy` infrastructure to load images that
  // encode their request headers in their pixels. Fun stuff!
  promise_test(() =>
    requestViaImage(
      "https://{{host}}:{{ports[https][0]}}/common/security-features/subresource/image.py")
    .then(result => {
        headers = result.headers;
        got = {
          "dest": headers["sec-fetch-dest"]
        };
        assert_header_dest_equals(got, "image");
      }),
    "Same-origin image");

  promise_test(() =>
    requestViaImage(
      "https://{{hosts[][www]}}:{{ports[https][0]}}/common/security-features/subresource/image.py")
    .then(result => {
        headers = result.headers;
        got = {
          "dest": headers["sec-fetch-dest"]
        };
        assert_header_dest_equals(got, "image");
      }),
    "Same-site image");

  promise_test(() =>
    requestViaImage(
      "https://{{hosts[alt][www]}}:{{ports[https][0]}}/common/security-features/subresource/image.py")
    .then(result => {
        headers = result.headers;
        got = {
          "dest": headers["sec-fetch-dest"]
        };
        assert_header_dest_equals(got, "image");
      }),
    "Cross-site image");
</script>
